+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 - **Package Handler** Fix: Trailing slash handling for compare tags URLs
 - **Package Handler** Fix: Handle find_all_elements for compare tags URL correctly between Github and GitLab
+- **Package Handler** Fix: Switched return statements in get_upgradable_packages to exit statements
 
 # 1.0.4 (2025-01-02)